Create a dedicated package for geo components
