+++ src/frontend/packages/geo-components/README.md
@@ -1,0 +1,16 @@
+# React-Admin components for geographic data display and input
+
+## Components
+
+
+
+## Commands
+
+### `npm run pre-publish`
+
+Build the library in CommonJS and ES mode.
+This command should always be run before publishing.
+
+### `npm run dev`
+
+Develop mode. Any change you make to a file in the `/src` directory will be immediately compiled.
