+++ src/frontend/packages/geo-components/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@semapps/geo-components",
+  "version": "0.1.18",
+  "description": "React-Admin component for geographic data display and input",
+  "license": "Apache-2.0",
+  "author": "Virtual Assembly",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.es.js",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "pre-publish": "rollup --config",
+    "dev": "rollup --watch --config"
+  },
+  "dependencies": {
+    "@material-ui/lab": "^4.0.0-alpha.57",
+    "autosuggest-highlight": "^3.1.1",
+    "lodash.throttle": "^4.1.1",
+    "react-leaflet": "^3.1.0",
+    "react-leaflet-markercluster": "^3.0.0-rc1"
+  },
+  "peerDependencies": {
+    "react": "^16.13.0",
+    "react-admin": "3.9.2",
+    "leaflet": "^1.7.1",
+    "leaflet.markercluster": "^1.4.1"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/preset-react": "^7.9.4",
+    "rollup": "^2.7.6",
+    "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-json": "^4.0.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-terser": "^5.3.0",
+    "rollup-pluginutils": "^2.8.2"
+  },
+  "gitHead": "4fa70690c5932b8c9d9cb5eb4c36a10f687703ff"
+}
