+++ src/frontend/packages/geo-components/src/index.js
@@ -1,0 +1,4 @@
+export { default as DefaultPopupContent } from './DefaultPopupContent';
+export { default as LocationInput } from './LocationInput';
+export { default as MapList } from './MapList';
+export { default as QueryStringUpdater } from './QueryStringUpdater';
